+++ src/containers/Search/Search.js
@@ -1,15 +1,15 @@
 import React, { Component } from 'react';
-import withStyles from '@material-ui/core/styles/withStyles';
+import { withStyles } from '@material-ui/core/styles';
 import { providers } from '../../utils';
 import SearchResults from '../SearchResults';
 import styles from './Search.styles';
 
 export class Search extends Component {
   render() {
-    const { onResultClick } = this.props;
+    const { onResultClick, renderSearchResult } = this.props;
 
     return (
-      <SearchResults onResultClick={onResultClick} />
+      <SearchResults onResultClick={onResultClick} renderSearchResult={renderSearchResult} />
     );
   }
 }